requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
djangorestframework_simplejwt
psycopg2>=2.7.5,<2.8.0
requests
flake8>=3.6.0,<3.7.0
gunicorn==20.0.4
whitenoise==5.2.0
dj-database-url==0.5.0